requirements.txt
flask
torch
diffusers
transformers
pillow